Add directory existence check

+++ install.sh
@@ -52,9 +52,10 @@ if [[ "$OSTYPE" = "darwin"* ]] ; then
 else
     vscode_config_path="$HOME/.config/Code/User/"
 fi
-
-ln -snvf "$(pwd)/Code/settings.json" "${vscode_config_path}/settings.json"
-ln -snvf "$(pwd)/Code/keybindings.json" "${vscode_config_path}/keybindings.json"
+if [ -e "${vscode_config_path}" ] ; then
+  ln -snvf "$(pwd)/Code/settings.json" "${vscode_config_path}/settings.json"
+  ln -snvf "$(pwd)/Code/keybindings.json" "${vscode_config_path}/keybindings.json"
+fi
 
 print_for_log2 "Finish installing dotfiles"
